refactor(#741): refactor entranceController
